refactor: Add EventHandler
