+++ Sources/OpenFeature/Client.swift
@@ -11,15 +11,4 @@ public protocol Client: Features {
     /// Hooks are run in the order they're added in the before stage. They are run in reverse order for all
     /// other stages.
     func addHooks(_ hooks: any Hook...)
-
-    /// Add a handler for a particular provider event
-    ///  - Parameter observer: The object observing the event.
-    ///  - Parameter selector: The selector to call for this event.
-    ///  - Parameter event: The event to listen for.
-    func addHandler(observer: Any, selector: Selector, event: ProviderEvent)
-
-    /// Remove a handler for a particular provider event
-    ///  - Parameter observer: The object observing the event.
-    ///  - Parameter event: The event being listened to.
-    func removeHandler(observer: Any, event: ProviderEvent)
 }
